+++ src/app/(dashboard)/dashboard/providers/page.tsx
@@ -34,7 +34,6 @@ import {
   loadProviderPageData,
 } from "./providerPageUtils";
 import type { ProviderEntry } from "./providerPageUtils";
-import { recordProviderNavigation, resolveHighlightedCard } from "./providerPageHighlightUtils";
 import {
   readProviderDisplayModePreference,
   shouldSyncProviderDisplayMode,
@@ -50,8 +49,7 @@ import AddCompatibleProviderModal from "./components/AddCompatibleProviderModal"
 import { CategoryDot } from "./components/CategoryDot";
 import { ImportProvidersFromFileModal } from "./components/ImportProvidersFromFileModal";
 import NoAuthProvidersSection from "./components/NoAuthProvidersSection";
-import ProviderCard from "./components/ProviderCard";
-import type { ProviderCardHandle } from "./components/ProviderCard";
+import HighlightableProviderCard from "./components/HighlightableProviderCard";
 import ProviderCountBadge from "./components/ProviderCountBadge";
 import ProviderSummaryCard from "./components/ProviderSummaryCard";
 import {
@@ -207,22 +205,6 @@ export default function ProvidersPage() {
   // #4240: media-category (serviceKind) filter — composes with activeCategory,
   // search and configured-only. null = no serviceKind filter.
   const [activeServiceKind, setActiveServiceKind] = useState<string | null>(null);
-
-  // a highlighted card is one that we should scroll into view and highlight upon navigation
-  const [highlightedProviderId, setHighlightedProviderId] = useState<string | null>(() => {
-    return history.state?.providerId ?? null;
-  });
-
-  const handleCardClick = useCallback((id: string) => {
-    recordProviderNavigation(id);
-  }, []);
-
-  const highlightedCardRef = useCallback(
-    (handle: ProviderCardHandle | null) => {
-      resolveHighlightedCard(handle, highlightedProviderId, () => setHighlightedProviderId(null));
-    },
-    [highlightedProviderId, setHighlightedProviderId]
-  );
   const notify = useNotificationStore();
   const sectionCategoryAliases: Record<string, string> = {
     cloud: "cloudagent",
@@ -927,7 +909,7 @@ export default function ProvidersPage() {
             data-testid="provider-compact-grid"
           >
             {compactProviderEntries.map((entry) => (
-              <ProviderCard
+              <HighlightableProviderCard
                 key={`compact-${entry.providerId}`}
                 providerId={entry.providerId}
                 provider={entry.provider}
@@ -936,9 +918,6 @@ export default function ProvidersPage() {
                 onToggle={(active) =>
                   handleToggleProvider(entry.providerId, entry.toggleAuthType, active)
                 }
-
-                onCardClick={handleCardClick}
-                ref={highlightedCardRef}
               />
             ))}
           </div>
@@ -1016,7 +995,7 @@ export default function ProvidersPage() {
                 <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                   {compatibleProviderEntries.map(
                     ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                      <ProviderCard
+                      <HighlightableProviderCard
                         key={providerId}
                         providerId={providerId}
                         provider={provider}
@@ -1025,9 +1004,6 @@ export default function ProvidersPage() {
                         onToggle={(active) =>
                           handleToggleProvider(providerId, toggleAuthType, active)
                         }
-
-                        onCardClick={handleCardClick}
-                        ref={highlightedCardRef}
                       />
                     )
                   )}
@@ -1093,7 +1069,7 @@ export default function ProvidersPage() {
                 {oauthProviderEntries
                   .filter((e) => !IDE_PROVIDER_IDS.has(e.providerId))
                   .map(({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={providerId}
                       providerId={providerId}
                       provider={provider}
@@ -1102,9 +1078,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   ))}
               </div>
@@ -1154,7 +1127,7 @@ export default function ProvidersPage() {
                 <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                   {ideProviderEntries.map(
                     ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                      <ProviderCard
+                      <HighlightableProviderCard
                         key={`ide-${providerId}`}
                         providerId={providerId}
                         provider={provider}
@@ -1163,9 +1136,6 @@ export default function ProvidersPage() {
                         onToggle={(active) =>
                           handleToggleProvider(providerId, toggleAuthType, active)
                         }
-
-                        onCardClick={handleCardClick}
-                        ref={highlightedCardRef}
                       />
                     )
                   )}
@@ -1207,16 +1177,13 @@ export default function ProvidersPage() {
               <p className="text-sm text-text-muted -mt-2">{t("webCookieProvidersDesc")}</p>
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {webCookieProviderEntries.map(({ providerId, provider, stats, toggleAuthType }) => (
-                  <ProviderCard
+                  <HighlightableProviderCard
                     key={providerId}
                     providerId={providerId}
                     provider={provider}
                     stats={stats}
                     authType="web-cookie"
                     onToggle={(active) => handleToggleProvider(providerId, toggleAuthType, active)}
-
-                    onCardClick={handleCardClick}
-                    ref={highlightedCardRef}
                   />
                 ))}
               </div>
@@ -1256,7 +1223,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {freeSectionEntries.map(
                   ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={`free-section-${providerId}`}
                       providerId={providerId}
                       provider={provider}
@@ -1265,9 +1232,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
@@ -1312,7 +1276,7 @@ export default function ProvidersPage() {
                   <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                     {llmProviderEntries.map(
                       ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                        <ProviderCard
+                        <HighlightableProviderCard
                           key={providerId}
                           providerId={providerId}
                           provider={provider}
@@ -1321,9 +1285,6 @@ export default function ProvidersPage() {
                           onToggle={(active) =>
                             handleToggleProvider(providerId, toggleAuthType, active)
                           }
-
-                          onCardClick={handleCardClick}
-                          ref={highlightedCardRef}
                         />
                       )
                     )}
@@ -1383,16 +1344,13 @@ export default function ProvidersPage() {
               <p className="text-sm text-text-muted -mt-2">{t("upstreamProxyProvidersDesc")}</p>
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {upstreamProxyEntries.map(({ providerId, provider, stats, toggleAuthType }) => (
-                  <ProviderCard
+                  <HighlightableProviderCard
                     key={providerId}
                     providerId={providerId}
                     provider={provider}
                     stats={stats}
                     authType="upstream-proxy"
                     onToggle={(active) => handleToggleProvider(providerId, toggleAuthType, active)}
-
-                    onCardClick={handleCardClick}
-                    ref={highlightedCardRef}
                   />
                 ))}
               </div>
@@ -1416,7 +1374,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {webFetchEntries.map(
                   ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={`webfetch-${providerId}`}
                       providerId={providerId}
                       provider={provider}
@@ -1425,9 +1383,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
@@ -1452,7 +1407,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {aggregatorProviderEntries.map(
                   ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={providerId}
                       providerId={providerId}
                       provider={provider}
@@ -1461,9 +1416,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
@@ -1488,7 +1440,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {enterpriseProviderEntries.map(
                   ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={providerId}
                       providerId={providerId}
                       provider={provider}
@@ -1497,9 +1449,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
@@ -1541,7 +1490,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {cloudAgentProviderEntries.map(
                   ({ providerId, provider, stats, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={providerId}
                       providerId={providerId}
                       provider={provider}
@@ -1550,9 +1499,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
@@ -1593,16 +1539,13 @@ export default function ProvidersPage() {
               <p className="text-sm text-text-muted -mt-2">{t("localProvidersDesc")}</p>
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {localProviderEntries.map(({ providerId, provider, stats, toggleAuthType }) => (
-                  <ProviderCard
+                  <HighlightableProviderCard
                     key={providerId}
                     providerId={providerId}
                     provider={provider}
                     stats={stats}
                     authType="local"
                     onToggle={(active) => handleToggleProvider(providerId, toggleAuthType, active)}
-
-                    onCardClick={handleCardClick}
-                    ref={highlightedCardRef}
                   />
                 ))}
               </div>
@@ -1642,16 +1585,13 @@ export default function ProvidersPage() {
               <p className="text-sm text-text-muted -mt-2">{t("searchProvidersDesc")}</p>
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {searchProviderEntries.map(({ providerId, provider, stats, toggleAuthType }) => (
-                  <ProviderCard
+                  <HighlightableProviderCard
                     key={providerId}
                     providerId={providerId}
                     provider={provider}
                     stats={stats}
                     authType="search"
                     onToggle={(active) => handleToggleProvider(providerId, toggleAuthType, active)}
-
-                    onCardClick={handleCardClick}
-                    ref={highlightedCardRef}
                   />
                 ))}
               </div>
@@ -1675,7 +1615,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {embeddingRerankProviderEntries.map(
                   ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={providerId}
                       providerId={providerId}
                       provider={provider}
@@ -1684,9 +1624,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
@@ -1711,7 +1648,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {imageProviderEntries.map(
                   ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={providerId}
                       providerId={providerId}
                       provider={provider}
@@ -1720,9 +1657,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
@@ -1763,16 +1697,13 @@ export default function ProvidersPage() {
               <p className="text-sm text-text-muted -mt-2">{t("audioProvidersDesc")}</p>
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {audioProviderEntries.map(({ providerId, provider, stats, toggleAuthType }) => (
-                  <ProviderCard
+                  <HighlightableProviderCard
                     key={providerId}
                     providerId={providerId}
                     provider={provider}
                     stats={stats}
                     authType="audio"
                     onToggle={(active) => handleToggleProvider(providerId, toggleAuthType, active)}
-
-                    onCardClick={handleCardClick}
-                    ref={highlightedCardRef}
                   />
                 ))}
               </div>
@@ -1796,7 +1727,7 @@ export default function ProvidersPage() {
               <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-4 gap-3">
                 {videoProviderEntries.map(
                   ({ providerId, provider, stats, displayAuthType, toggleAuthType }) => (
-                    <ProviderCard
+                    <HighlightableProviderCard
                       key={providerId}
                       providerId={providerId}
                       provider={provider}
@@ -1805,9 +1736,6 @@ export default function ProvidersPage() {
                       onToggle={(active) =>
                         handleToggleProvider(providerId, toggleAuthType, active)
                       }
-
-                      onCardClick={handleCardClick}
-                      ref={highlightedCardRef}
                     />
                   )
                 )}
